Weapon added and fixed. Check Description
It took me a little while to find out what was  the issue. 
The code for shooting originally used the `get_parent()` to get the world class, spawn the bullet based on the parent's `global_position`.
However, before the refactoring, the player's parent was the world scene. That's why it worked, it got the player's global position based on the world scene.
In the rework, using a weapon scene as a child from player, had the player AS the parent. Thus, it got the player's position, not the world's position. So now it's using get_tree().get_root(), which calls the root instead